NewsApp/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.2
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
newsapi-python==0.2.7
requests==2.31.0
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.18.1
